mpqp-generator/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-WTF==1.2.2
psycopg2-binary==2.9.10
SQLAlchemy==2.0.36
Werkzeug==3.1.3
gunicorn==23.0.0
PyMuPDF==1.25.3
python-docx==1.1.2
openpyxl==3.1.5
chromadb==0.6.3
